% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogsci-package.R
\docType{package}
\name{bcogsci-package}
\alias{bcogsci}
\alias{bcogsci-package}
\title{bcogsci: Data and models for the book "An Introduction to Bayesian Data Analysis for Cognitive Science"}
\description{
Datasets and models included in the book "An Introduction to Bayesian Data Analysis for Cognitive Science".
}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{b.nicenboim@tilburguniversity.edu} (\href{https://orcid.org/0000-0002-5176-3943}{ORCID})

Authors:
\itemize{
  \item Daniel Schad \email{danieljschad@gmail.com} (\href{https://orcid.org/ORCID-ID}{ORCID})
  \item Shravan Vasishth \email{vasishth@uni-potsdam.de} (\href{https://orcid.org/ORCID-ID}{ORCID})
}

}
\keyword{internal}
