% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{compute_meansd_parent}
\alias{compute_meansd_parent}
\title{Compute Mean and Standard Deviation of the Parent Distribution of a Truncated Distribution}
\usage{
compute_meansd_parent(
  a = -Inf,
  b = Inf,
  mean_trunc = NULL,
  sd_trunc = NULL,
  startvals = c(1, 1)
)
}
\arguments{
\item{a}{Lower bound of the truncated distribution. Default is `-Inf`.}

\item{b}{Upper bound of the truncated distribution. Default is `Inf`.}

\item{mean_trunc}{The mean of the truncated distribution.}

\item{sd_trunc}{The standard deviation of the truncated distribution.}

\item{startvals}{Starting values for the optimization algorithm. Default is c(1, 1).}
}
\value{
The location and scale parameters of the parent distribution.
}
\description{
This function computes the mean and standard deviation of the parent distribution
of a truncated distribution.
}
\examples{
\dontrun{
compute_meansd_parent(a=0, b=1000, mean_trunc=500, sd_trunc=50)
}
}
