% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ab}
\alias{df_ab}
\title{Experimental condition where the target was detected from the replication of Experiment 2 of Raymond, Shapiro, and Arnell (1992), The Attentional Blink by \insertCite{grassi_two_2021;textual}{bcogsci}}
\format{
A data frame with 2,101 rows and 4 variables:
\describe{
\item{subj}{Subject index.}
\item{trial}{Trial number}
\item{lag}{Lag of the trial. When the probe is absent, lag is equal to "NA".}
\item{probe_correct}{Response accuracy on the probe (i.e., the letter X). 1 correct, 0 incorrect.}
}
}
\usage{
df_ab
}
\description{
The dataset is from the UK participants of the replication of Experiment 2 of  \insertCite{raymond1992temporary}{bcogsci}. In this experiment, participants are presented with a fast stream of letters. Letters are black except for a white one (the target). In half of the trials, the stream includes the letter "X" (the probe). The probe can appear in the target position or lag up to eight relative serial positions after the target. In the control condition, the participants are asked to detect the probe (i.e., the letter "X"). In the experimental condition, the participants first have to report the identity of the target and then indicate the presence or absence of the probe (see Raymond et al. 1992, Fig. 1). For this experiment, the refresh rate of the screen was 60Hz.
}
\details{
Complete dataset can be found in [df_ab_complete].Original data set can be found in https://figshare.com/articles/dataset/Two_replications_of_Raymond_Shapiro_and_Arnell_1992_The_Attentional_Blink/6834353
}
\references{
\insertAllCited{}
}
\keyword{datasets}
