% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_blp_complete}
\alias{df_blp_complete}
\title{Data from the lexical decision task  of the British Lexicon project \insertCite{keuleers2012british}{bcogsci}}
\format{
A data frame with 2240802 rows and 8 variables:
\describe{
\item{subj}{The subject id.}
\item{block}{Block number.}
\item{lex}{lexicality of the string: Either word or non-word.}
\item{trial}{Trial number.}
\item{string}{String of letters.}
\item{acc}{1 for correct, 0 for incorrect.}
\item{rt}{Raw reaction times of all strings in milliseconds. (Observations with zero reaction times are omited)}
\item{freq}{Frequency per million words, see details.}
}
}
\usage{
df_blp_complete
}
\description{
Reaction times (in milliseconds) and accuracy for the lexical decision task of the British Lexicon project \insertCite{keuleers2012british}{bcogsci}.
}
\details{
The column names are modified from the original dataset (see below). The freq column indicates the frequency per million words according to the British National Corpus (BNC, available at http://www.natcorp.ox.ac.uk), with the following modifications: Non-words have a frequency of zero, and all words have a frequency of at least 0.01. The original data set can be found in http://crr.ugent.be/programs-data/lexicon-projects.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
