% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_buerki}
\alias{df_buerki}
\title{Estimates (mean effect and their standard error) reported in a study on the picture-word interference paradigm \insertCite{BuerkiEtAl2020;textual}{bcogsci}.}
\format{
A data frame with 162 rows and 4 variables:
\describe{
\item{study}{The publication and experiment details of each study.}
\item{d}{The interference effect in milliseconds.}
\item{se}{Standard error of the effect in milliseconds.}
\item{study_id}{A numerical id for each study.}
}
}
\usage{
df_buerki
}
\description{
Data from a meta-analysis by \insertCite{BuerkiEtAl2020;textual}{bcogsci}.
}
\details{
The data provide, for each individual study that investigated the picture-word
 interference phenomenon, the estimated interference effect and its
 standard error in milliseconds.

The data and associated code are available from https://osf.io/k6f4c/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
