% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_chineseRCs}
\alias{df_chineseRCs}
\title{Data from reading studies that list the difference in reading times in milliseconds object vs. subject relative clauses in Chinese. \insertCite{VasishthMScStatistics;textual}{bcogsci}}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{study.id}{The study id.}
\item{study}{The study name.}
\item{y}{The difference in object vs. subject reading times at the critical region (the relative clause head noun) in milliseconds.}
\item{se}{The standard error of the difference in means between object vs. subject relative clauses.}
}
}
\usage{
df_chineseRCs
}
\description{
The data-set is from a meta-analysis reported in \insertCite{VasishthMScStatistics;textual}{bcogsci}; the replication attempt was reported in \insertCite{VasishthMScStatistics;textual}{bcogsci}, and contains reading times from a self-paced reading experiment on subject and object relative clauses in Chinese. There are 40 subjects who read 15 items in a Latin square design. The estimates are from the relative clause head noun, which is the critical region of interest.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
