% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_eeg}
\alias{df_eeg}
\title{EEG averages in the N400 spatio-temporal window at the noun region of  \insertCite{nieuwlandLargescaleReplicationStudy2018;textual}{bcogsci} for one lab.}
\format{
A data frame with 2,863 rows and 6 variables:
\describe{
  \item{subj}{Subject id.}
  \item{cloze}{Cloze probability (from 0 to 1)}
  \item{item}{Item number}
  \item{n400}{Average N400 signal at the noun.}
  \item{cloze_ans}{Number of answers for the noun (Cloze prob. multiplied by 44)}
  \item{N}{Number of answers, 44}
}
}
\usage{
df_eeg
}
\description{
EEG averages in the N400 spatio-temporal window at the noun region together with
the noun Cloze probability for the participants  of the Edinburgh's lab in
\insertCite{nieuwlandLargescaleReplicationStudy2018;textual}{bcogsci}. Participants
were presented with sentences word by word, and the EEG signal was recorded.
}
\details{
\insertCite{nieuwlandLargescaleReplicationStudy2018;textual}{bcogsci} presented a
replication attempt of an original experiment of
\insertCite{delongProbabilisticWordPreactivation2005;textual}{bcogsci}
with sentences like the following

 a. The day was breezy so the boy went outside to fly a kite.
 b. The day was breezy so the boy went outside to fly an airplane.

The papers focus on the article (an, an), this dataset has the EEG averages
and cloze probability of the *nouns* (kite/airplane)

Complete dataset can be found in [df_eeg_complete]. Original data and stimuli can be
found in public_noun_data.txt (https://osf.io/q7dsk/) and
Stimuli.zip (https://osf.io/acu4r/)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
