% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fallacy}
\alias{df_fallacy}
\title{Responses for the conjunction fallacy or Linda problem Experiment \insertCite{TverskyKahneman1983}{bcogsci}}
\format{
A data frame with 268 rows and 2 variables:
\describe{
\item{source}{Whether the data is from the university or Mechanical Turk.}
\item{answer}{0 indicates option a, 1 indicates option b}
}
}
\usage{
df_fallacy
}
\description{
Responses from the replication of \insertCite{Paolaccietal;textual}{bcogsci} of the conjunction fallacy experiment or Linda problem \insertCite{TverskyKahneman1983}{bcogsci}. The dataset is adapted (and simplified) from https://osf.io/wzsf5/
}
\details{
\insertCite{Paolaccietal;textual}{bcogsci} examined whether the results of some classic experiments differ between a university pool population and participants recruited from Mechanical Turk among them the conjunction fallacy.

The conjunction fallacy shows that people often fail to regard a combination of events as less probable than a single event in the combination. In the experiment a text as the following appears followed

*Linda is 31 years old, single, outspoken, and very bright. She majored in philosophy. As a student, she was deeply concerned with issues of discrimination and social justice, and also participated in anti-nuclear demonstrations.*

*Which is more probable?*

a. *Linda is a bank teller.*
b. *Linda is a bank teller and is active in the feminist movement.*

The majority of those asked chose option b. Even if it's less probable since Pr(a & b) is smaller or equal than Pr(b).

In this dataset `0` indicates option "a" and `1` option `b`.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
