% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_gg05_complete}
\alias{df_gg05_complete}
\title{Reading times of the self-paced reading experiment 1 conducted by  \insertCite{grodnergibson2005;textual}{bcogsci}.}
\format{
A data frame with 10,331 rows and 8 variables:
\describe{
 \item{subj}{}
 \item{item}{}
\item{condition}{ Either object-RC (objgap) or subject-RC (subjgap).}
\item{word_position}{Word number in the sentence.}
\item{region_position}{Region position (see example above)}
\item{RT}{Reading times in milliseconds}
\item{residRT}{Residual reading times computed by subtracting variance due to length from unadjusted reading times (using the procedure described by Frazier & Clifton, 1986).}
\item{qcorrect}{Whether the comprehension question was answered correctly or not.}
}
}
\usage{
df_gg05_complete
}
\description{
Reading times of every region of Experiment 1 of \insertCite{grodnergibson2005;textual}{bcogsci}.
}
\details{
The experiment examined sentences with object-RC (relative clauses) vs sentences with subject-RC, such as the following (regions are indicated with |):

- (SRC) The reporter | who | the photographer | sent | to the editor | hoped for a story.
- (ORC) The reporter | who | sent | the photographer | to the editor | hoped for a story.

The is dataset is a slightly modified version (namely, column names were changed, and a an error in word position column was fixed) of the original dataset. The complete description of the experiments included in the dataset can be found in \insertCite{grodnergibson2005;textual}{bcogsci}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
