% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_indiv}
\alias{df_indiv}
\title{Working memory capacity scores and speed of the rapid automatized naming task from the confirmatory data of \insertCite{NicenboimEtAlCogSci2018;textual}{bcogsci}}
\format{
A data frame with 82 rows and 5 variables:
\describe{
\item{subj}{The subject id.}
\item{mean_rspeed}{Mean speed at the RAN.}
\item{se_rspeed}{SE of the speed at the RAN.}
\item{mean_pcu}{Mean pcu at the operation span task}
\item{se_rspeed}{SE of the speed at the operation span task.}
}
}
\usage{
df_indiv
}
\description{
Data from from the confirmatory experiment  of \insertCite{NicenboimEtAlCogSci2018;textual}{bcogsci}.
The data shows (1) partial-credit unit scores, which indicate the mean proportion of correctly recalled items within the sets of an operation span task, as an index of working memory capacity \insertCite{conway2005working}{bcogsci} together with their SE, and (2) the mean speeds in the rapid automatized naming task, RAN, \insertCite{DenklaRudel1976}{bcogsci} as a measure of reading fluency (in characters/second) together with their SE.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
