% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_lagoE1}
\alias{df_lagoE1}
\title{Data from a self-paced reading study on Spanish reported in \insertCite{lago2015agreement}{bcogsci}}
\format{
A data frame with 768 rows and 6 variables:
\describe{
\item{subj}{The subject id.}
\item{item}{The item id.}
\item{rt}{Reading time at the post-critical word in the sentence.}
\item{int}{The interference conditions (high and low).}
\item{x}{Sum contrast coding of the interference conditions.}
\item{expt}{The experiment id.}
}
}
\usage{
df_lagoE1
}
\description{
Reading times (in milliseconds) from a self-paced reading study on Spanish reported in \insertCite{lago2015agreement}{bcogsci}.
}
\details{
The column names are modified from the original dataset. The reading times are from the post-critical word (en) in the sentences shown below.

-(a) Ungram , singular attractor (interference condition)
La *nota* que la chica escribieron en la clase alegró a su amiga
The note that the girl wrote-pl during class cheered her friend up
-(b) Ungram , plural attractor (baseline condition)
Las *notas* que la chica escribieron en la clase alegraron a su amiga
The notes that the girl wrote-pl during class cheered her friend up

There are 32 subjects who read 48 items in a Latin square design. The data shown here focus on two conditions from the experiment, high and low interference conditions in a subject-verb number agreement configuration in ungrammatical sentences. The language under study is Spanish.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
