% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pupil_complete}
\alias{df_pupil_complete}
\title{Average pupil size for all the participant of the control experiment of \insertCite{wahnPupilSizesScale2016;textual}{bcogsci}.}
\format{
A data frame with 2,228 rows and 4 variables:
\describe{
  \item{subj}{Subject id.}
  \item{trial}{Trial number.}
  \item{p_size}{Pupil size.}
  \item{load}{Load condition.}
}
}
\usage{
df_pupil_complete
}
\description{
The data contain trial averaged pupil size (after 100 ms) for the
participants of the experiment from
\insertCite{wahnPupilSizesScale2016}{bcogsci}.
In this experiment, a participant covertly tracked between
zero and five objects among several randomly moving objects
on a computer screen. This task is called  multiple object tracking
or MOT \insertCite{pylyshynTrackingMultipleIndependent1988}{bcogsci} task. First,
several objects appear on the screen, and a subset of them are
indicated as "targets" at the beginning. Then, the objects start
moving randomly across the screen and become indistinguishable. After
several seconds, the objects stop moving and the participant need to
indicate which objects were the targets.
Notice that this experiment was conducted with the Eyelink-II
eyetracker which measures the pupils in arbitrary units
\insertCite{hayesMappingCorrectingInfluence2016}{bcogsci}.
}
\details{
Full dataset can be found in the file `PLOS_raw_pupildata_controlexperiment.csv` https://osf.io/qtzjb/
}
\references{
\insertAllCited{}
}
\keyword{datasets}
