% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pupil_pilot}
\alias{df_pupil_pilot}
\title{Average pupil size for one participant sampled every 10ms of the control experiment of \insertCite{wahnPupilSizesScale2016;textual}{bcogsci}.}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{time}{Time in milliseconds}
  \item{p_size}{Pupil size}
  \item{att_load}{Load condition}
}
}
\usage{
df_pupil_pilot
}
\description{
The data contain 10 samples of the average pupil size for
participant 701 under no load condition of the experiment from
\insertCite{wahnPupilSizesScale2016}{bcogsci}.
Notice that this experiment was conducted with the Eyelink-II
eyetracker which measures the pupils in arbitrary units
\insertCite{hayesMappingCorrectingInfluence2016}{bcogsci}.
}
\details{
Full dataset can be found in the file `PLOS_raw_pupildata_controlexperiment.csv` https://osf.io/qtzjb/
}
\references{
\insertAllCited{}
}
\keyword{datasets}
