% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_recall}
\alias{df_recall}
\title{Free recall data for one  participant of Experiment 1 of \insertCite{oberauerWorkingMemoryCapacity2019;textual}{bcogsci}.}
\format{
A data frame with 92 rows and 7 variables:
\describe{
  \item{subj}{Subject id.}
  \item{set_size}{Set size: 2, 4, 6, or 8 elements.}
  \item{correct}{Correct or not.}
  \item{trial}{Trial number.}
  \item{block}{Block number.}
  \item{session}{Session number.}
}
}
\usage{
df_recall
}
\description{
Subset of the data of one participant (subj = 10) from Experiment 1 of
\insertCite{oberauerWorkingMemoryCapacity2019;textual}{bcogsci}.
The participant was
presented word lists of varying lengths, and then was asked to recall a word
given its position on the list, and type the probed word.
(Trials with multiple forced choices were removed).
}
\details{
Complete dataset can be found in [df_recall_complete]. Original data can be
found in PairsRSS1_all.dat in https://osf.io/qy5sd/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
