% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_recall_complete}
\alias{df_recall_complete}
\title{Recall data for participants of Experiment 1 of \insertCite{oberauerWorkingMemoryCapacity2019;textual}{bcogsci}.}
\format{A data frame with 92 rows and 7 variables:
\describe{
  \item{subj}{Subject id.}
  \item{session}{Session number.}
  \item{block}{Block number.}
  \item{trial}{Trial number.}
  \item{set_size}{Set size: 2, 4, 6, or 8 elements.}
  \item{response_size_list}{Number of options in the multiple choice task, 0 indicates free recall (by typing).}
  \item{response_size_new_words}{Number of lures (non presented words) in the multiple choice task.}
  \item{response_category}{Type of answer: either correct, intrusion, or new.}
  \item{correct}{Correct or not.}
}}
\usage{
df_recall_complete
}
\description{
Complete data of participants from Experiment 1 of
\insertCite{oberauerWorkingMemoryCapacity2019;textual}{bcogsci}. Participants
were presented word lists of varying lengths, and then was asked to recall a
 word given its position on the list.
}
\details{
Original data can be
found in PairsRSS1_all.dat in https://osf.io/qy5sd/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
