% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_red}
\alias{df_red}
\title{Data from an experiment on the effect of fertility status on the color of clothing worn by women, originally reported by \insertCite{beall2013women;textual}{bcogsci}}
\format{
A data frame with 124 rows and 5 variables:
\describe{
\item{risk}{A 0,1 variable indicating whether the participant is at risk of pregnancy or not.}
\item{age}{The age of the participant in years.}
\item{red}{A 0,1 coded variable indicating whether the participant reported wearing red.}
\item{pink}{A 0,1 coded variable indicating whether the participant reported wearing pink.}
\item{redorpink}{A 0,1 coded variable indicating whether the participant reported wearing either red or pink.}
}
}
\usage{
df_red
}
\description{
The data set is from a study by \insertCite{beall2013women;textual}{bcogsci}, and contains information 
about the color of the clothing worn (red, pink, or redorpink) when the participant (female) is at risk of becoming pregnant (is ovulating, self-reported).
}
\references{
\insertAllCited{}
}
\keyword{datasets}
