% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sbi}
\alias{df_sbi}
\title{Mean reading times of interference effects in subject-verb dependencies in the experiments of the meta analysis of \insertCite{JaegerEngelmannVasishth2017;textual}{bcogsci}.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{publication}{}
\item{effect}{Estimated mean effect in milliseconds}
\item{SE}{Standard error of the effect}
}
}
\usage{
df_sbi
}
\description{
Mean reading times of interference effects in subject-verb dependencies in the experiments of the meta analysis of \insertCite{JaegerEngelmannVasishth2017;textual}{bcogsci}. According to similarity-based interference accounts when two nouns are more similar to each other, there is greater processing difficulty (i.e., longer reading times at a critical region). The data contains 12 estimates.
}
\details{
Complete dataset can be found in [df_sbi_complete].
}
\references{
\insertAllCited{}
}
\keyword{datasets}
