% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sbi_complete}
\alias{df_sbi_complete}
\title{Mean reading times of interference/facilitation effects in the experiments of the meta analysis of \insertCite{JaegerEngelmannVasishth2017;textual}{bcogsci}.}
\format{
A data frame with 3058 rows and 4 variables:
\describe{
 \item{publication}{}
 \item{language}{EN = English, CN = Chinese, HI = Hindi, FR = French, SP = Spanish, AR = Arabic}
 \item{method_measure}{Either SPR_RT = Reading times from self-paced reading, or ET_FPRT, first pass reading times from eye-tracking while reading}
 \item{cue}{}
 \item{int_type}{Type of interference, either pro = proactive, or retro = retroactive}
 \item{distr_pos}{Indicates where the distractors are in the sentence}
\item{effect}{Estimated mean effect in milliseconds}
\item{SE}{Standard error of the effect}
\item{target_type}{Match or Mismatch}
\item{dep_type}{Type of the dependency investigated: relf = reflexive-antecedent, reci =  reciprocal-antecedent, agreement = subject-verb/agreement attraction, nonagreement = subject-verb}
\item{verb_type}{}
\item{prominence1}{}
\item{prominence2}{}
}
}
\usage{
df_sbi_complete
}
\description{
Mean reading times of interference effects in subject-verb dependencies and reflexive/reciprocal-antecedent dependencies in the experiments of the meta analysis of \insertCite{JaegerEngelmannVasishth2017;textual}{bcogsci}. According to similarity-based interference accounts when two nouns are more similar to each other, there is greater processing difficulty (i.e., longer reading times at a critical region). The data contains 77 estimates from twelve experiments.
}
\details{
The is dataset is a slightly modified version (namely, column names were changed) of the dataset that can be found in https://osf.io/du3qp. The complete description of the experiments included in the dataset can be found in https://osf.io/4ehyw/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
