% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_schizophrenia}
\alias{df_schizophrenia}
\title{Data from a simple visual tracking experiment for schizophrenic patients and controls from \insertCite{belin1990analysis;textual}{bcogsci}}
\format{
A data frame with 372 rows and 4 variables:
\describe{
\item{subj}{The subject id.}
\item{trial}{Trial number}
\item{patient}{Indicates whether the it is a schizophrenic patient (1) or not (0).}
\item{rt}{Reaction times in milliseconds.}
}
}
\usage{
df_schizophrenia
}
\description{
Reaction times (in milliseconds) for 11 non-schizophrenics and 6
schizophrenics (30 measurements for each person) doing a simple visual tracking experiment from \insertCite{belin1990analysis;textual}{bcogsci}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
