% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_source_monitoring}
\alias{df_source_monitoring}
\title{Response frequencies of 128 participants who took part in a source-monitoring experiment in \insertCite{wickelmaier2018using;textual}{bcogsci}.}
\format{
A data frame with 384 rows and 8 variables:
\describe{
\item{experiment}{ write-say or think-say}
\item{age}{ Age of the respondent in years.}
\item{gender}{ Gender of the respondent.}
\item{subj}{ Subject id.}
\item{source}{ Item source, a, b or b (new)}
\item{a`, `b`, `N}{ Number of responses for each type of stimuli}
}
}
\usage{
df_source_monitoring
}
\description{
This dataset is from a source-monitoring experiment and was originally provided by the psychotools package \insertCite{psychotools}{bcogsci}. In this version the format of the original data frame was modified for ease of use with Stan.
}
\details{
In this type of experiment, subjects study items from (at least) two different sources, A and B. After the presentation of the study items, subjects are required to classify each item as coming from source A, B, or as new: N (that is, a distractor). In the version of the experiment from \insertCite{wickelmaier2018using;textual}{bcogsci}, participants had to read items either quietly (think) or aloud (say). In the recall task, they wrote them down (write) or read them aloud (say).
}
\references{
\insertAllCited{}
}
\keyword{datasets}
