% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_spacebar}
\alias{df_spacebar}
\title{Reaction times in milliseconds from a participant repeatedly pressing the space bar.}
\format{A data frame with 361 rows and 2 variables:
\describe{
  \item{rt}{Reaction time in milliseconds}
  \item{trial}{Trial number}
}}
\source{
\url{}
}
\usage{
df_spacebar
}
\description{
Data from a participant (author) repeatedly pressing the space bar
as fast as possible, without paying attention to any stimuli.
The data are reaction times in milliseconds.
}
\keyword{datasets}
