% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_swets08}
\alias{df_swets08}
\title{Data from a self-paced reading experiment that records reading times in milliseconds at the post-critical region. \insertCite{swets2008underspecification;textual}{bcogsci}}
\format{
A data frame with 5,184 rows and 6 variables:
\describe{
\item{subj}{The subject id.}
\item{item}{The item id.}
\item{resp.RT}{Response times to questions.}
\item{qtype}{The three levels of the between-subjects factor, question type.}
\item{attachment}{The three levels of the within-subjects factor, attachment type.}
\item{RT}{Reading times at the post-critical region.}
}
}
\usage{
df_swets08
}
\description{
The dataset is from a self-paced reading experiment by \insertCite{swets2008underspecification;textual}{bcogsci}, and contains reading times from a 3x3 design.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
