% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_VOTmandarin}
\alias{df_VOTmandarin}
\title{Voice onset times (VOTs) from a Mandarin experiment conducted by  \insertCite{li2013effect;textual}{bcogsci}.}
\format{
A data frame with 20 rows and 6 variables:
\describe{
 \item{subject}{Subject id, 10 female speakers and 10 male speakers.}
 \item{meanVOT}{Mean voice onset time for each speaker in milliseconds.}
\item{meanvdur}{Mean vowel duration for each speaker in milliseconds.}
\item{sevdur}{Standard error of the mean vowel duration in milliseconds.}
\item{sevot}{Standard error of the mean voice onset time in milliseconds.}
\item{c_meanvdur}{Centered mean vowel duration.}
}
}
\usage{
df_VOTmandarin
}
\description{
Voice onset times from Mandarin speakers in an experiment reported in \insertCite{li2013effect;textual}{bcogsci}.
}
\details{
The experiment has mean voice onset times and their standard errors, along with mean vowel durations and their standard error. 

The data and associated code are available from https://osf.io/g4zpv/.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
