% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{ginv2}
\alias{ginv2}
\title{Compute the Generalized Inverse with Nicely Formatted Output}
\usage{
ginv2(x)
}
\arguments{
\item{x}{A numeric matrix for which the generalized inverse is to be computed.}
}
\value{
A matrix that is the generalized inverse of the input matrix \code{x}, with dimension names preserved and entries formatted as fractions.
}
\description{
This function computes the generalized inverse of a matrix using the \code{\link[MASS]{ginv}} function from the \code{MASS} package. The output is nicely formatted with dimension names preserved and the matrix entries represented as fractions.
}
\examples{
\dontrun{
A <- matrix(c(1, 2, 3, 4), 2, 2)
ginv2(A)
}

}
