% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{normal_predictive_distribution}
\alias{normal_predictive_distribution}
\title{Generate Prior Predictive Distributions}
\usage{
normal_predictive_distribution(mu_samples, sigma_samples, N_obs)
}
\arguments{
\item{mu_samples}{A numeric vector of samples from the prior distribution of the mean parameter (\eqn{\mu}).}

\item{sigma_samples}{A numeric vector of samples from the prior distribution of the standard deviation parameter (\eqn{\sigma}).}

\item{N_obs}{An integer indicating the number of observations to generate for each sample.}
}
\value{
A tibble containing the generated data sets. Each row represents a generated observation with the columns:
\describe{
  \item{trialn}{The trial number (from 1 to \code{N_obs}).}
  \item{t_pred}{The predicted value generated from the normal distribution with parameters \code{mu} and \code{sigma}.}
  \item{iter}{The iteration number corresponding to the sample from the prior distributions.}
}
}
\description{
This function generates prior predictive distributions by taking samples from the priors
and using them to generate a data set based on the normal distribution. The slower but easier to read function shown in section 3.3 is \link{normal_predictive_distribution_slow}.
}
\examples{
# Example usage:
mu_samples <- rnorm(100, 0, 1)  # 100 samples from a normal prior with mean 0 and standard deviation 1
sigma_samples <- runif(100, 0.1, 2)  # 100 samples from a uniform prior with values between 0.1 and 2
N_obs <- 10  # Number of observations to generate for each sample
prior_predictive_data <- normal_predictive_distribution(mu_samples, sigma_samples, N_obs)
head(prior_predictive_data)

}
