% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{qpf}
\alias{qpf}
\title{Quantile Probability Plot Data Preparation}
\usage{
qpf(df_grouped, quantiles = c(0.1, 0.3, 0.5, 0.7, 0.9))
}
\arguments{
\item{df_grouped}{A data frame grouped by an experimental condition, containing columns
`rt` (response time) and `acc` (accuracy, with `0` indicating incorrect responses and `1` indicating correct responses).}

\item{quantiles}{A numeric vector specifying the quantiles to be calculated for the response time distribution.
Default is `c(0.1, 0.3, 0.5, 0.7, 0.9)`.}
}
\value{
A data frame with columns for quantiles of response times (`rt_q`), probabilities of correct
and incorrect responses (`p`), quantile labels (`q`), and response type (`response`), structured for plotting.
The resulting data frame contains one row for each quantile for both correct and incorrect responses
across the specified experimental conditions.
}
\description{
This function prepares data for a quantile probability plot, which visualizes the relationship
between response times and accuracy across experimental conditions.
}
\details{
The function calculates the specified quantiles of response times for correct and incorrect responses
by each experimental condition. It also calculates the proportion of correct and incorrect responses by condition,
repeating this proportion for each quantile to align with the quantile data. The resulting data structure is ideal
for generating a quantile probability plot.
}
\examples{
# Assuming `df_blp` is a data frame containing columns `lex` (condition), `rt` (response time), and `acc` (accuracy)
df_blp_lex_q <- df_blp \%>\%
  group_by(lex) \%>\%
  qpf()

}
