% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-binomial_cloze}
\alias{stan-binomial_cloze}
\title{Stan model for fitting a binomial likelihood with a beta prior.}
\description{
Stan model for fitting a binomial likelihood with a beta prior.
}
\examples{
# The file can be located as follows:
binomial <- system.file("stan_models", "binomial_cloze.stan", package = "bcogsci")
# See the models code:
file.show(binomial)
}
