% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-normal}
\alias{stan-normal}
\title{Stan model for fitting a normal likelihood.}
\description{
Stan model for fitting a normal likelihood.
}
\examples{
# The file can be located as follows:
normal <- system.file("stan_models", "normal.stan", package = "bcogsci")
# See the models code:
file.show(normal)
}
