% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-normal2}
\alias{stan-normal2}
\title{Stan model for fitting a normal likelihood with a for-loop.}
\description{
Stan model for fitting a normal likelihood with a for-loop.
}
\examples{
# The file can be located as follows:
normal <- system.file("stan_models", "normal2.stan", package = "bcogsci")
# See the models code:
file.show(normal2)
}
