% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-pupil_gen}
\alias{stan-pupil_gen}
\title{Stan model for fitting a linear regression to the pupil size data and generating predictive distributions.}
\description{
Stan model for fitting a linear regression to the pupil size data and generating predictive distributions.
}
\examples{
# The file can be located as follows:
normal <- system.file("stan_models", "pupil_gen.stan", package = "bcogsci")
# See the models code:
file.show(pupil_gen)
}
