% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-pupil_model}
\alias{stan-pupil_model}
\title{Stan model for fitting a linear regression to the pupil size data.}
\description{
Stan model for fitting a linear regression to the pupil size data.
}
\examples{
# The file can be located as follows:
normal <- system.file("stan_models", "pupil_model.stan", package = "bcogsci")
# See the models code:
file.show(pupil_model)
}
