% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_models.R
\name{stan-rema}
\alias{stan-rema}
\title{Stan model for a random-effects meta-analysis.}
\description{
Stan model for a random-effects meta-analysis.
}
\examples{
# The file can be located as follows:
normal <- system.file("stan_models", "rema.stan", package = "bcogsci")
# See the model code:
file.show(rema)
}
