% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_powerpose}
\alias{df_powerpose}
\title{Data from an experiment on the effect of power posing on testosterone levels, originally reported by \insertCite{carney2010power;textual}{bcogsci}}
\format{
A data frame with 39 rows and 4 variables:
\describe{
\item{id}{Subject id.}
\item{item}{The item id, ranging from 1 to 16. Each subject id is preceded by an i.}
\item{hptreat}{Power pose treatment; high power pose or low power pose.}
\item{testm1}{Testosterone level before treatment.}
\item{testm1}{Testosterone level after treatment.}
}
}
\usage{
df_powerpose
}
\description{
The dataset is from a study by \insertCite{carney2010power;textual}{bcogsci}, and contains testosterone measurements before and after treatement (which can be either high or low power posing). There is one set of measurements (before and after) from each subject.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
