% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_spacebar}
\alias{df_spacebar}
\title{Finger tapping times in milliseconds from a participant repeatedly pressing the space bar.}
\format{
A data frame with 361 rows and 2 variables:
\describe{
  \item{t}{Finger tapping time in milliseconds}
  \item{trial}{Trial number}
}
}
\source{
\url{}
}
\usage{
df_spacebar
}
\description{
Data from a participant (first author) repeatedly tapping the space bar
as fast as possible. The data are finger tapping times in milliseconds and trial number.
}
\keyword{datasets}
